+++ app/components/Game/GameBlock.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import gameImage from '../../images/game-image.jpg';
 
-const GameBlock = styled.div`
+ export const GameBlock = styled.div`
   flex: 1;
   border: 1px solid red;
   class: game-block;
@@ -10,4 +10,4 @@ const GameBlock = styled.div`
   background-size: cover;
   background-position: center;
 `;
-export default GameBlock;
+
