--- app/components/GameComponent/index.js
@@ -1,3 +1,0 @@
-// finish import and exports
-
-import GameBlock from 'GameBlock';
