+++ app/components/Info/Arrows.js
@@ -1,0 +1,3 @@
+export const ArrowComponent = styled.li`
+  border: 1px solid green;
+`;
