+++ app/components/Info/LineCounter.js
@@ -1,0 +1,3 @@
+export const LineComponent = styled.li`
+  border: 1px solid green;
+`;
