+++ app/components/Info/NextTetromino.js
@@ -1,0 +1,3 @@
+export const NextComponent = styled.li`
+  border: 1px solid green;
+`;
