+++ app/components/Info/Score.js
@@ -1,0 +1,3 @@
+export const Score = styled.li`
+  border: 1px solid green;
+`;
