+++ app/components/Info/StartGame.js
@@ -1,0 +1,6 @@
+export const StartGame = styled.button`
+  border: 1px solid green;
+  background-color: green;
+  display: flex;
+  flex: 1;
+`;
