--- app/components/InfoComponent/index.js
@@ -1,1 +1,0 @@
-// finish import and exports
