+++ app/containers/App/index.js
@@ -15,7 +15,7 @@ import NotFoundPage from 'containers/NotFoundPage/Loadable';
 
 import GlobalStyle from '../../global-styles';
 
-export default function App() {
+export function App() {
   return (
     <div>
       <Switch>
